removed goal display component from List file, more modular. Relates #89
